+++ src/TodoForm.jsx
@@ -1,0 +1,11 @@
+function TodoForm(){
+    return (
+        <>
+        <label htmlFor="todoTitle">Todo &nbsp;</label>
+        <input id="todoTitle"></input>
+        <button>Add Todo</button>
+        </>
+    )
+}
+
+export default TodoForm
